keep building libzmq as an Extension for Windows PyPy wheels
delvewheel + binary libzmq doesn't seem to work for PyPy (not sure why)

PyPy itself claims to require separate installation of vcredist, so we do, too,
instead of trying to bundle it.
